pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["Jhonatan-Diniz <jhonatandiniz360@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.22.1"}
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
tzdata = "^2024.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"

[tool.taskipy.tasks]
dev = 'fastapi dev main.py --port 8087'
run = 'uvicorn main:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
